完成入库单增删改查，正在编码确认

+++ src/main/java/com/gqj/entity/MaterialBillDetail.java
@@ -1,65 +1,65 @@
 package com.gqj.entity;
 
-import java.math.BigDecimal;
-
 public class MaterialBillDetail {
-    private Long detailId;
+	private Long detailId;
 
-    private Long baseToolId;
+	private Long baseToolId;
 
-    private Long billId;
+	private Long billId;
 
-    private BigDecimal detailBillAmount;
+	private Double detailBillAmount;
 
-    private String detailRemark;
+	private String detailRemark;
 
-    private Long posId;
+	private Long posId;
 
-    public Long getDetailId() {
-        return detailId;
-    }
+	public Long getDetailId() {
+		return detailId;
+	}
 
-    public void setDetailId(Long detailId) {
-        this.detailId = detailId;
-    }
+	public void setDetailId(Long detailId) {
+		this.detailId = detailId;
+	}
 
-    public Long getBaseToolId() {
-        return baseToolId;
-    }
+	public Long getBaseToolId() {
+		return baseToolId;
+	}
 
-    public void setBaseToolId(Long baseToolId) {
-        this.baseToolId = baseToolId;
-    }
+	public void setBaseToolId(Long baseToolId) {
+		this.baseToolId = baseToolId;
+	}
 
-    public Long getBillId() {
-        return billId;
-    }
+	public Long getBillId() {
+		return billId;
+	}
 
-    public void setBillId(Long billId) {
-        this.billId = billId;
-    }
+	public void setBillId(Long billId) {
+		this.billId = billId;
+	}
 
-    public BigDecimal getDetailBillAmount() {
-        return detailBillAmount;
-    }
+	public Double getDetailBillAmount() {
+		return detailBillAmount;
+	}
 
-    public void setDetailBillAmount(BigDecimal detailBillAmount) {
-        this.detailBillAmount = detailBillAmount;
-    }
+	public void setDetailBillAmount(
+			Double detailBillAmount) {
+		this.detailBillAmount = detailBillAmount;
+	}
 
-    public String getDetailRemark() {
-        return detailRemark;
-    }
+	public String getDetailRemark() {
+		return detailRemark;
+	}
 
-    public void setDetailRemark(String detailRemark) {
-        this.detailRemark = detailRemark == null ? null : detailRemark.trim();
-    }
+	public void setDetailRemark(String detailRemark) {
+		this.detailRemark = detailRemark == null ? null
+				: detailRemark.trim();
+	}
 
-    public Long getPosId() {
-        return posId;
-    }
+	public Long getPosId() {
+		return posId;
+	}
 
-    public void setPosId(Long posId) {
-        this.posId = posId;
-    }
+	public void setPosId(Long posId) {
+		this.posId = posId;
+	}
 }
